adjusted authorization for adventure asset at handler level; unified adventure repositories; complemented unit tests

+++ src/main/java/me/moirai/discordbot/core/application/usecase/adventure/DeleteAdventureHandler.java
@@ -3,20 +3,23 @@ package me.moirai.discordbot.core.application.usecase.adventure;
 import org.apache.commons.lang3.StringUtils;
 
 import me.moirai.discordbot.common.annotation.UseCaseHandler;
+import me.moirai.discordbot.common.exception.AssetAccessDeniedException;
 import me.moirai.discordbot.common.exception.AssetNotFoundException;
 import me.moirai.discordbot.common.usecases.AbstractUseCaseHandler;
 import me.moirai.discordbot.core.application.usecase.adventure.request.DeleteAdventure;
-import me.moirai.discordbot.core.domain.adventure.AdventureDomainRepository;
+import me.moirai.discordbot.core.domain.adventure.Adventure;
+import me.moirai.discordbot.core.domain.adventure.AdventureRepository;
 
 @UseCaseHandler
 public class DeleteAdventureHandler extends AbstractUseCaseHandler<DeleteAdventure, Void> {
 
+    private static final String USER_NO_PERMISSION = "User does not have permission to delete adventure";
     private static final String ADVENTURE_NOT_FOUND = "Adventure to be deleted was not found";
     private static final String ID_CANNOT_BE_NULL_OR_EMPTY = "Adventure ID cannot be null or empty";
 
-    private final AdventureDomainRepository repository;
+    private final AdventureRepository repository;
 
-    public DeleteAdventureHandler(AdventureDomainRepository repository) {
+    public DeleteAdventureHandler(AdventureRepository repository) {
         this.repository = repository;
     }
 
@@ -31,8 +34,12 @@ public class DeleteAdventureHandler extends AbstractUseCaseHandler<DeleteAdventu
     @Override
     public Void execute(DeleteAdventure command) {
 
-        repository.findById(command.getId())
+        Adventure adventure = repository.findById(command.getId())
                 .orElseThrow(() -> new AssetNotFoundException(ADVENTURE_NOT_FOUND));
+
+        if (!adventure.canUserWrite(command.getRequesterDiscordId())) {
+            throw new AssetAccessDeniedException(USER_NO_PERMISSION);
+        }
 
         repository.deleteById(command.getId());
